[GraphBolt] Rename and move utils files. (#6627)
