+++ python/dgl/graphbolt/internal/__init__.py
@@ -1,5 +1,5 @@
 """Utility functions for GraphBolt."""
-from .internal import *
+from .utils import *
 from .sample_utils import *
 from .datapipe_utils import *
 from .item_sampler_utils import *
