+++ python/dgl/graphbolt/subgraph_sampler.py
@@ -6,8 +6,8 @@ from typing import Dict
 from torch.utils.data import functional_datapipe
 
 from .base import etype_str_to_tuple
+from .internal import unique_and_compact
 from .minibatch_transformer import MiniBatchTransformer
-from .utils import unique_and_compact
 
 __all__ = [
     "SubgraphSampler",
